save progress working on adding traits
